Convert Forecast to Model and develop Epoch
A massive change with no real incremental steps.
The RecommendationGroup is now an Epoch and shows up in many tables as new
sort of partition key. All references to the RecommendationGroup or group_id
are now Epoch or epoch_id and added to new record creation where possible.

The notion of a Forecast is now one of several possible types of Model.
References to this for the database are updated and the API is modified.
Additional changes may follow.