Add read/write scope to database interactions
Major re-write to core database connection mechanism.
We want to allow many read-replica databases and a main write-database.
Database/API calls that merely read information should target the read-replicas.

This commit implements this using a `DatabaseConnection` manager that lazily
access `db.read` and `db.write` sessions based on scopes defined in the config.

There will likely be numerous issues to come with these scopes. For now,
this commit assumes the read and write scope point to the same session.
In the future, we need to ensure operations throughput the project are
consistently using the correct scope (and passing it into entity methods).