Refactor project into separate packages under namespace
This changes the struction of the project by splitting out the single
monolithic package source into refitt-core and others. For now, the
bulk of the project is temporarily in refitt-main; this will change in
follow-up commits gradually.

Some issues with tests exist in this commit because of assumptions
made on imports between unit tests and integration tests.